Revert "refactor: moved config to one place"
This reverts commit 6e318ff225636f3f22bd2227c8b2adcf76a16a4e.
