+++ frontend/config/config.tsx
@@ -1,0 +1,10 @@
+namespace Config {
+
+    /**
+     * URL to the Backend Server - Will be replaced automatically by the CD 
+     */ 
+    export const SERVER_URL = 'http://localhost:3000';
+
+}
+
+export default Config;
